requirements.txt
django-crispy-forms==1.4.0
django-registration-redux==1.1
django-profiles==0.2